Move dcp_install to submodule init.
