+++ bifrost/dcp/Job.py
@@ -204,50 +204,6 @@ class Job:
 
         return module_encoded
 
-    def dcp_install(self):
-
-        # install and initialize dcp-client
-
-        # TODO: Call this before python_deploy is run, to ensure that dcp-client is available
-        # -> Do NOT move this back to Job.__init__; that location precludes custom scheduler settings
-        # -> This is now being called in compute_for, after Job.__init__, but before returning the Class
-
-        from bifrost import node, npm
-
-        def _parse_version(version_string):
-            version_list = version_string.split('.')
-            version = tuple(int(version_element) for version_element in version_list)
-            return version
-
-        def _npm_checker(package_name):
-
-            npm_io = io.StringIO()
-            with contextlib.redirect_stdout(npm_io):
-                npm.list_modules(package_name)
-            npm_check = npm_io.getvalue()
-
-            if '(empty)' in npm_check:
-                print('installing ' + package_name)
-                npm.install(package_name)
-            else:
-                try:
-                    package_latest = npm.package_latest_version(package_name)
-                    package_current = npm.package_current_version(package_name)
-                    if _parse_version(package_current) < _parse_version(package_latest):
-                        print('installing version ' + package_latest + ' of ' + package_name)
-                        npm.install(package_name + '@' + package_latest)
-                    else:
-                        print('proceeding with currently installed version of ' + package_name)
-                except ValueError:
-                    print('installing default npm version of ' + package_name)
-                    npm.install(package_name)
-
-        _npm_checker('dcp-client')
-
-        node.run("""
-        if ( !globalThis.dcpClient ) globalThis.dcpClient = require("dcp-client").init(scheduler);
-        """, { 'scheduler': self.scheduler })
-
     def __dcp_run(self):
 
         from bifrost import node
